backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg[binary]>=3.1.0
python-dotenv==1.0.1
dj-database-url==2.1.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
whiteNoise==6.6.0


 